Tidy tests.
